use the radio tower icon for default broadcast image
